Add a non root user (muslrust)
This is a workaround the the (far to common) permission issues that
usually arises when using docker images as build environments. Instead
of running and building as root we now run as the muslrust user that has
been added which gets the user-id and group-id of 1000:1000 that in
many cases should also correspond to the default user for Linux users.

It might not cover all cases and therefore there is also build arguments
to configure the id of the user being created. Sadly it requires the
users of the docker image to build their own version but at least it is
doable to get around the issue.